DAY-5 added scheduler, apiClient, apiService

+++ src/main/java/com/kodilla/rental/domain/dto/api/nbp/RatesDto.java
@@ -13,5 +13,5 @@ import lombok.NoArgsConstructor;
 public class RatesDto {
 
     @JsonProperty
-    private float mid;
+    private double mid;
 }
